Refactor code to avoid errors

+++ app/helpers/users_helper.rb
@@ -4,14 +4,10 @@ module UsersHelper
   end
 
   def current_user_links(user)
-    if actual_user?
-      link_to 'Edit User Account', edit_user_path(user), class: 'btn btn-light'
-    end
+    link_to 'Edit User Account', edit_user_path(user), class: 'btn btn-light' if actual_user?
   end
 
   def current_user_links2(user)
-    if actual_user?
-      link_to 'Delete Account', user_path(user), method: :delete, class: 'btn btn-light ml-2'
-    end
+    link_to 'Delete Account', user_path(user), method: :delete, class: 'btn btn-light ml-2' if actual_user?
   end
 end
